data/requirements.txt
dnspython==2.7.0
Faker==30.1.0
mysql-connector-python==9.0.0
numpy==2.1.2
pandas==2.2.3
psycopg2-binary==2.9.9
pymongo==4.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2